chore(room schema): refactor room schema query - create another file to accomadate the long room schema queries
[Finishes #165207900]
