+++ schema.py
@@ -3,7 +3,9 @@ import api.location.schema
 import api.block.schema
 import api.floor.schema
 import api.room.schema
-import api.room.schema_query
+import api.room.remote_room_schema
+import api.room.analytics_schema
+import api.room.events_schema
 import api.room_resource.schema
 import api.role.schema
 import api.user.schema
@@ -23,7 +25,9 @@ class Query(
     api.location.schema.Query,
     api.block.schema.Query,
     api.floor.schema.Query,
-    api.room.schema_query.Query,
+    api.room.remote_room_schema.Query,
+    api.room.analytics_schema.Query,
+    api.room.events_schema.Query,
     api.room_resource.schema.Query,
     api.role.schema.Query,
     api.user.schema.Query,
