+++ tests/test_rooms/test_query_rooms.py
@@ -31,7 +31,8 @@ class QueryRooms(BaseTestCase):
         expected_response = paginated_rooms_response
         self.assertEqual(paginate_query, expected_response)
 
-    @patch("api.room.schema_query.get_google_api_calendar_list", spec=True,
+    @patch("api.room.remote_room_schema.get_google_api_calendar_list",
+           spec=True,
            return_value=get_calendar_list_mock_data())
     def test_query_remote_rooms(self, mock_get_json):
         CommonTestCases.admin_token_assert_in(
